Restructure Vue project
Rerun `npm create vite@latest`
