+++ index.html
@@ -2,13 +2,13 @@
 <html lang="en">
 <head>
     <meta charset="UTF-8">
-    <link rel="icon" href="/favicon.ico" />
+    <link rel="icon" href="/favicon.ico">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@400;500&display=swap" rel="stylesheet">
     <title>LineageOS Downloads</title>
 </head>
 <body>
     <div id="app"></div>
-    <script type="module" src="/src/js/main.js"></script>
+    <script type="module" src="/src/main.js"></script>
 </body>
 </html>
