+++ package.json
@@ -1,22 +1,31 @@
 {
   "name": "updater-client",
   "version": "0.0.0",
+  "private": true,
+  "type": "module",
   "scripts": {
     "dev": "vite",
-    "build": "vite build"
+    "build": "vite build",
+    "preview": "vite preview",
+    "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs --fix --ignore-path .gitignore",
+    "format": "prettier --write src/"
   },
   "dependencies": {
-    "@mdi/font": "^5.9.55",
-    "@vitejs/plugin-vue": "^4.6.0",
+    "@mdi/font": "^7.4.47",
     "axios": "^1.6.3",
     "node-forge": "^1.3.1",
-    "simplebar": "^5.3.9",
-    "vite": "^4.5.1",
-    "vue": "^3.0.0-rc.1",
+    "simplebar": "^6.2.5",
+    "vue": "^3.3.11",
     "vue-router": "^4.2.5",
     "vuex": "^4.1.0"
   },
   "devDependencies": {
-    "@vue/compiler-sfc": "^3.0.0-rc.1"
+    "@rushstack/eslint-patch": "^1.3.3",
+    "@vitejs/plugin-vue": "^4.5.2",
+    "@vue/eslint-config-prettier": "^8.0.0",
+    "eslint": "^8.49.0",
+    "eslint-plugin-vue": "^9.17.0",
+    "prettier": "^3.0.3",
+    "vite": "^5.0.10"
   }
 }
