+++ src/js/ApiService.js
@@ -1,4 +1,4 @@
-import store from './store';
+import store from '../store';
 import axios from 'axios';
 import {API_HOSTNAME} from './config';
